update (add README)

+++ process/delete_actor_process.php
@@ -5,12 +5,20 @@ require_once '../classes/Notification.php';
 require_once '../classes/Actor.php';
 
 if ($_GET['id'] == 0 || !isset($_GET)) {
-    header('Location: ../index.php');
+    header('Location: ../admin/upload.php');
     exit;
 }
 
 $id = $_GET['id'];
 $actor = Actor::getActor($id, Actor::getActors());
+
+if (
+    !unlink('../uploads/actors/' . $actor['picture_actors']) ||
+    !unlink('../uploads/actors/picture_round/' . $actor['picture_round_actors'])
+) {
+    header('Location: ../admin/upload.php?error=' . Notification::ERROR_DELETE_ACTOR);
+    exit;
+}
 
 try {
     $sql = 'DELETE FROM actors_series WHERE id_actors = :id';
@@ -21,11 +29,8 @@ try {
     $query = Database::getConnection()->prepare($sql);
     $query->execute(['id' => $id]);
 
-    unlink('../uploads/actors/' . $actor['picture_actors']);
-    unlink('../uploads/actors/picture_round/' . $actor['picture_round_actors']);
-
     header('Location: ../admin/upload.php?succes=' . Notification::DELETE_ACTOR);
     exit;
 } catch (PDOException $e) {
-    header('Location: ../admin/upload.php?error='. Notification::ERROR_DELETE_ACTOR);
+    header('Location: ../admin/upload.php?error=' . Notification::ERROR_DELETE_ACTOR);
 }
