+++ process/delete_serie_process.php
@@ -1,10 +1,10 @@
-<?php 
+<?php
 
 require_once "../classes/Database.php";
 require_once "../classes/Notification.php";
 
 if (!isset($_POST['series'])) {
-    header('Location: ../index.php');
+    header('Location: ../admin/index.php');
     exit;
 }
 
@@ -21,7 +21,10 @@ try {
     $query->execute(['id' => $id_series]);
     $series = $query->fetch();
 
-    unlink('../uploads/series/' . $series['picture_series']);
+    if (!unlink('../uploads/series/' . $series['picture_series'])) {
+        header('Location: ../admin/upload.php?error=' . Notification::ERROR_DELETE_SERIE);
+        exit;
+    }
 
     $sql = 'DELETE FROM series WHERE id_series = :series';
     $query = Database::getConnection()->prepare($sql);
